Se corrigio codigo de home y se elimino la carpeta desktop
