+++ src/components/screens/camara/styles.module.scss
@@ -1,0 +1,69 @@
+@import '../../../styles/main.scss';
+
+.containerCamara {
+  width: 100%;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  align-items: center;
+  font-family: 'Open Sans', sans-serif;
+}
+
+.topbar {
+  width: 100%;
+  height: 10%;
+  background-color: $color-blue;
+  display: flex;
+  justify-content: space-between;
+}
+
+.iconBack {
+  width: 70%;
+  margin-top: 10px;
+  height: 60px;
+}
+
+.button {
+  background-color: $color-blue;
+  border: 0;
+  margin: 0;
+  margin-left: 2vh;
+  width: 80px;
+  height: 70px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.iconLogo {
+  width: 40%;
+  height: 84px;
+  display: flex;
+  justify-content: center;
+  margin-left: 30%;
+}
+
+.cam {
+  width: 100%;
+  height: 106vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  align-items: center;
+}
+
+.iconCamara {
+  height: 100%;
+  width: 100%;
+}
+
+.footer {
+  visibility: hidden;
+}
+
+@media all and (max-width: 760px) {
+  .topbar {
+    visibility: hidden;
+  }
+}
