--- src/components/screens/desktop/foundations-selector/contenedor-fundaciones/cards/styles.module.scss
@@ -1,56 +1,0 @@
-@import '../../../../../../styles/main.scss';
-
-.containerCards {
-  width: 200px;
-  height: 210px;
-  border: solid 4px;
-  border-color: $border-target;
-  border-radius: 20px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 10px;
-  position: relative;
-}
-
-.containterBackground {
-  width: 80vh;
-  height: 100%;
-}
-
-.checkbox {
-  position: absolute;
-  margin-left: 64px;
-  margin-top: 80px;
-  transform: scale(2);
-  padding: 60px;
-}
-
-.containerTitle {
-  width: 90%;
-  height: 25%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  justify-content: space-around;
-}
-
-.titleFoundation {
-  font-size: 20px;
-  font-weight: 500;
-  text-align: center;
-}
-
-.containerDates {
-  margin-top: 30px;
-  margin-left: 10px;
-  font-size: 12px;
-  font-family: Arial, Helvetica, sans-serif;
-  line-height: 6px;
-}
-
-.titleDate {
-  font-size: 16px;
-  font-weight: 600;
-}
