--- src/components/screens/desktop/foundations-selector/contenedor-fundaciones/styles.module.scss
@@ -1,18 +1,0 @@
-@import '../../../../../styles/main.scss';
-
-.contCards {
-  width: 940px;
-  height: 510px;
-  display: flex;
-  overflow: auto;
-  flex-wrap: wrap;
-  justify-content: center;
-  align-items: center;
-}
-
-.titleContainer {
-  margin-bottom: 80px;
-  font-size: 18px;
-  color: $color-title;
-  font-family: Arial, Helvetica, sans-serif;
-}
