--- src/components/screens/desktop/scanner/styles.module.scss
@@ -1,165 +1,0 @@
-@import '../../../../styles/main.scss';
-
-.containerScannerMain {
-  width: 100%;
-  height: 96vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  align-items: center;
-  font-family: 'Open Sans', sans-serif;
-  font-style: normal;
-}
-
-.containerText {
-  width: 80%;
-  height: 52vh;
-  display: flex;
-  flex-direction: column;
-  margin-left: 18vh;
-  font-size: 22px;
-  line-height: 12px;
-}
-
-.titleEnd {
-  color: $color-blue;
-  font-size: 46px;
-  font-weight: bold;
-  margin-top: 6vh;
-}
-
-.subtitle {
-  color: $color-blue;
-  font-style: italic;
-  font-weight: bold;
-  font-size: 26px;
-}
-
-.text {
-  font-weight: 600;
-  color: $color-text-end;
-  line-height: 24px;
-  font-size: 20px;
-}
-
-.containterScanner {
-  width: 100%;
-  height: 40vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  align-items: center;
-}
-
-.titleScan {
-  color: $color-blue;
-  width: 34%;
-  height: 50vh;
-  font-size: 30px;
-  font-weight: bold;
-  text-align: center;
-}
-
-.containerCard {
-  width: 250px;
-  height: 260px;
-  border: solid 4px $border-target;
-  border-radius: 10px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 0;
-  position: relative;
-}
-
-.scanner {
-  width: 150px;
-  height: 150px;
-  position: relative;
-}
-
-.footer {
-  visibility: hidden;
-}
-
-@media all and (max-width: 770px) {
-  .containerScannerMain {
-    height: 90vh;
-  }
-
-  .containerText {
-    height: 40vh;
-    width: 68%;
-    line-height: 4px;
-    margin-left: 0;
-    justify-content: center;
-  }
-
-  .titleEnd {
-    font-size: 24px;
-    margin-top: 4vh;
-    margin-bottom: 2vh;
-  }
-
-  .subtitle {
-    font-size: 18px;
-    width: 100%;
-    height: 28px;
-    line-height: 30px;
-  }
-
-  .text {
-    width: 80%;
-    font-size: 14px;
-    line-height: 20px;
-  }
-
-  .titleScan {
-    font-size: 20px;
-    height: 8vh;
-    width: 50%;
-    margin-bottom: -30px;
-  }
-
-  .containterScanner {
-    transform: translateY(40px);
-  }
-
-  .containerCard {
-    width: 170px;
-    height: 180px;
-    box-shadow: none;
-  }
-
-  .scanner {
-    width: 100px;
-    height: 100px;
-
-  }
-
-  .footer {
-    visibility: visible;
-    height: 2vh;
-    margin-bottom: -20px;
-    width: 80%;
-    justify-content: center;
-  }
-
-  .hrFooter {
-    width: 60%;
-    background-color: $hr-color;
-    height: 1px;
-  }
-
-  .textFooter {
-    font-size: 8px;
-    color: $copy-color;
-    height: 30px;
-    line-height: 14px;
-    font-family: Open Sans;
-    font-style: normal;
-    font-weight: normal;
-    text-align: center;
-  }
-}
