--- src/components/screens/desktop/tablet/styles.module.scss
@@ -1,25 +1,0 @@
-@import '../../../../styles/main.scss';
-
-.containerTabletMain {
-  width: 100%;
-  height: 98vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  align-items: center;
-  font-family: 'Open Sans', sans-serif;
-  font-style: normal;
-}
-
-.containerTablet {
-  width: 100%;
-  height: 90vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  align-items: center;
-}
-
-.footer {
-  visibility: hidden;
-}
