+++ src/components/screens/foundations-selector/contenedor-fundaciones/styles.module.scss
@@ -1,0 +1,18 @@
+@import '../../../../styles/main.scss';
+
+.contCards {
+  width: 940px;
+  height: 510px;
+  display: flex;
+  overflow: auto;
+  flex-wrap: wrap;
+  justify-content: center;
+  align-items: center;
+}
+
+.titleContainer {
+  margin-bottom: 80px;
+  font-size: 18px;
+  color: $color-title;
+  font-family: Arial, Helvetica, sans-serif;
+}
