+++ src/components/screens/foundations-selector/styles.module.scss
