+++ src/components/screens/home/styles.module.scss
@@ -1,0 +1,190 @@
+@import '../../../styles/main.scss';
+
+.containerHome {
+  width: 100%;
+  height: 90vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  align-items: center;
+  font-family: 'Open Sans', sans-serif;
+}
+
+.containerWelcome {
+  width: 90%;
+  height: 65%;
+  display: flex;
+  flex-direction: column;
+  line-height: 2px;
+  margin-top: 40vh;
+}
+
+.titleHome {
+  color: $color-blue;
+  font-size: 90px;
+  font-style: normal;
+  font-weight: bold;
+}
+
+.textHome {
+  font-size: 30px;
+  width: 90%;
+  font-weight: bold;
+  color: $black-color;
+}
+
+.hr {
+  display: block;
+  margin-top: 8px;
+  margin-bottom: 8px;
+  background-color: $hr-color;
+  height: 2px;
+  width: 90%;
+}
+
+.containerButton {
+  width: 90%;
+  height: 25%;
+  margin-top: calc(2% + 28px);
+}
+
+.button1 {
+  background-color: $pink-color;
+  color: $text-button-color;
+  font-size: 20px;
+  font-family: Open Sans;
+  font-style: normal;
+  font-weight: bold;
+  border-radius: 8px;
+  margin-right: 24px;
+  border-color: $pink-color;
+  width: 260px;
+  height: 64px;
+  transition: all 0.1s;
+
+  &:hover {
+    transform: translateY(-4px);
+    cursor: pointer;
+  }
+}
+
+.button2 {
+  background-color: $color-blue;
+  color: $text-button-color;
+  font-size: 20px;
+  font-family: Open Sans;
+  font-style: normal;
+  font-weight: bold;
+  border-radius: 8px;
+  border-color: $color-blue;
+  width: 260px;
+  height: 64px;
+  transition: all 0.1s;
+
+  &:hover {
+    transform: translateY(-4px);
+    cursor: pointer;
+  }
+}
+
+.footer {
+  visibility: hidden;
+}
+
+@media all and (max-width: 640px) {
+
+  /*Aca cuando uso Eslint me hace un salto de linea de la llave hacia el containerButton y queda un espacio*/
+  .containerHome {
+    height: 100vh;
+  }
+
+  .containerWelcome {
+    height: 50vh;
+    line-height: 4px;
+    margin-top: 8px;
+    text-align: center;
+    justify-content: space-evenly;
+  }
+
+  .titleHome {
+    width: 60%;
+    font-size: 66px;
+    margin-top: 10vh;
+    margin-left: 4vh;
+  }
+
+  .textHome {
+    font-size: 26px;
+    box-sizing: content-box;
+    width: 60%;
+    text-align: start;
+    line-height: 30px;
+    margin-left: 8vh;
+    margin-top: 0;
+    margin-bottom: 0;
+  }
+
+  .hr {
+    width: 60%;
+    margin-bottom: 30px;
+  }
+
+  .containerButton {
+    display: flex;
+    align-content: center;
+    height: 50vh;
+    width: 60%;
+    flex-direction: column;
+    line-height: 4px;
+    text-align: center;
+    justify-content: center;
+  }
+
+  .button1 {
+    width: 100%;
+    height: 90px;
+    margin-bottom: 10px;
+    transition: all 0.1s;
+    box-sizing: content-box;
+
+    &:hover {
+      transform: translateY(-4px);
+      cursor: pointer;
+    }
+  }
+
+  .button2 {
+    width: 100%;
+    height: 90px;
+    margin-bottom: 90px;
+    transition: all 0.1s;
+    box-sizing: content-box;
+    align-items: center;
+
+    &:hover {
+      transform: translateY(-4px);
+      cursor: pointer;
+    }
+  }
+
+  .footer {
+    visibility: visible;
+    height: 10vh;
+  }
+
+  .hrFooter {
+    width: 60%;
+    background-color: $hr-color;
+    height: 1px;
+  }
+
+  .textFooter {
+    font-size: 10px;
+    color: $copy-color;
+    height: 30px;
+    line-height: 14px;
+    font-family: Open Sans;
+    font-style: normal;
+    font-weight: normal;
+  }
+}
