+++ src/components/screens/medicine-selector/medicine-container/cards/styles.module.scss
