+++ src/components/screens/medicine-selector/medicine-container/index.js
