+++ src/components/screens/medicine-selector/medicine-container/styles.module.scss
@@ -1,0 +1,20 @@
+@import '../../../../styles/main.scss';
+
+.contCards {
+  width: 60%;
+  height: 510px;
+  display: flex;
+  overflow: auto;
+  flex-wrap: wrap;
+  justify-content: center;
+  align-items: center;
+  margin-left: 20%;
+}
+
+.titleContainer {
+  margin-bottom: 80px;
+  font-size: 18px;
+  color: $color-title;
+  font-family: Arial, Helvetica, sans-serif;
+  margin-left: 26%;
+}
