+++ src/components/screens/medicine-selector/styles.module.scss
