+++ src/components/screens/tablet/container-card/card/slider/styles.module.scss
@@ -1,0 +1,50 @@
+@import '../../../../../../styles/main.scss';
+
+.inputRange {
+  background: $color-blue;
+  position: relative;
+  width: 80%;
+  height: 8px;
+  margin-left: 4vh;
+  margin-top: 32px;
+  border-radius: 40px;
+}
+
+.trackBackground {
+  height: 10px;
+  left: 0;
+  position: absolute;
+  right: 0;
+  border-radius: 60px;
+}
+
+.cursor {
+  position: absolute;
+  left: 50px;   /*este margin mueve la posicion del cursor*/
+  transition: left 0.3s ease-out;
+  height: 10px;
+
+}
+
+.labelValue {
+  position: absolute;
+  top: -1.8rem;
+}
+
+.slider {
+  background: $color-blue;
+  border: 2px solid $color-blue;
+  border-radius: 80%;
+  cursor: pointer;
+  display: block;
+  height: 14px;
+  margin-top: 24px;
+  position: absolute;
+  width: 1rem;
+  transition: all;
+
+  &:active { /*por ahora va este codigo pero la idea es que al hacer ondrag cambie el color de la barra*/
+    transform: translateX(20vh);
+    cursor: pointer;
+  }
+}
