+++ src/index.main.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 
-import Home from './components/screens/desktop/home';
-import MedicineSelector from './components/screens/desktop/medicine-selector';
-import Tablet from './components/screens/desktop/tablet';
-import FoundationsSelector from './components/screens/desktop/foundations-selector';
-import Scanner from './components/screens/desktop/scanner';
-import Camara from './components/screens/desktop/camara';
-import Confirmation from './components/screens/desktop/confirmation';
+import Home from './components/screens/home';
+import MedicineSelector from './components/screens/medicine-selector';
+import Tablet from './components/screens/tablet';
+import FoundationsSelector from './components/screens/foundations-selector';
+import Scanner from './components/screens/scanner';
+import Camara from './components/screens/camara';
+import Confirmation from './components/screens/confirmation';
 
 class Main extends React.Component {
   state = {
